reorganize events
