--- burner/keyring/lib/starknet.js
@@ -1,5 +1,0 @@
-import { Provider } from "starknet/provider";
-
-export const provider = new Provider({
-  sequencer: { network: "goerli-alpha" },
-});
